fix: add book input handling

+++ mini-project-2/src/main/java/com/stratpoint/projects/AudioBook.java
@@ -2,9 +2,9 @@ package com.stratpoint.projects;
 
 public class AudioBook extends Book{
     private String type;
-    private double lengthInMinutes;
+    private int lengthInMinutes;
 
-    public AudioBook(String title, String author, String genre, String ISBN, double length ) {
+    public AudioBook(String title, String author, String genre, String ISBN, int length ) {
         super(title, author, genre, ISBN);
         this.type = "Audio";
         this.lengthInMinutes = length;
